refactoring day02
