+++ 2019/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod day01;
+pub mod day02;
